// src/API.ts
export enum DriverRegistrationStatus {
  NEW = "NEW",
  COMPLETE = "COMPLETE",
  REJECTED = "REJECTED",
}
// ...
export type UpdateDriverInput = {
  id: string,
  name?: string | null,
  email?: string | null,
  image?: string | null,
  phone_number?: string | null,
  rating?: number | null,
  inviteCode?: string | null,
  lng?: string | null,
  lat?: string | null,
  verificationStatus?: DriverRegistrationStatus | null,
  onlineStatus?: boolean | null,
  fcmToken?: string | null,
  _version?: number | null,
};
// ...
export enum ModelSortDirection {
  ASC = "ASC",
  DESC = "DESC",
}
// ...
export type UpdateDriverMutationVariables = {
  input: UpdateDriverInput,
  condition?: ModelDriverConditionInput | null,
};

export type UpdateDriverMutation = {
  updateDriver?:  {
    __typename: "Driver",
    id: string,
    name: string,
    email: string,
    image?: string | null,
    phone_number?: string | null,
    rating?: number | null,
    inviteCode?: string | null,
    lng?: string | null,
    lat?: string | null,
    verificationStatus?: DriverRegistrationStatus | null,
    onlineStatus?: boolean | null,
    Cars?:  {
      __typename: "ModelCarConnection",
      items:  Array< {
        __typename: "Car",
        id: string,
        model: string,
        make: string,
        year: number,
        plate_number: string,
        color: string,
        driverID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    Ratings?:  {
      __typename: "ModelRatingConnection",
      items:  Array< {
        __typename: "Rating",
        id: string,
        rating?: number | null,
        comment?: string | null,
        driver_image?: string | null,
        driver_name?: string | null,
        customer_name?: string | null,
        customer_image?: string | null,
        tripsID: string,
        driverID: string,
        userID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    DriverWallets?:  {
      __typename: "ModelDriverWalletConnection",
      items:  Array< {
        __typename: "DriverWallet",
        id: string,
        amount: number,
        driverID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    LicenseDocuments?:  {
      __typename: "ModelLicenseDocumentConnection",
      items:  Array< {
        __typename: "LicenseDocument",
        id: string,
        image?: string | null,
        name?: string | null,
        expiry_date?: string | null,
        driverID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    RegistrationDocuments?:  {
      __typename: "ModelRegistrationDocumentConnection",
      items:  Array< {
        __typename: "RegistrationDocument",
        id: string,
        image?: string | null,
        name?: string | null,
        expiry_date?: string | null,
        driverID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    InsuranceDocuments?:  {
      __typename: "ModelInsuranceDocumentConnection",
      items:  Array< {
        __typename: "InsuranceDocument",
        id: string,
        image?: string | null,
        name?: string | null,
        expiry_date?: string | null,
        driverID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    fcmToken?: string | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};
// ...
export type RatingsByDriverIDQueryVariables = {
  driverID: string,
  sortDirection?: ModelSortDirection | null,
  filter?: ModelRatingFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type RatingsByDriverIDQuery = {
  ratingsByDriverID?:  {
    __typename: "ModelRatingConnection",
    items:  Array< {
      __typename: "Rating",
      id: string,
      rating?: number | null,
      comment?: string | null,
      driver_image?: string | null,
      driver_name?: string | null,
      customer_name?: string | null,
      customer_image?: string | null,
      tripsID: string,
      driverID: string,
      userID: string,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
    } | null >,
    nextToken?: string | null,
    startedAt?: number | null,
  } | null,
};
// ...
export type ListTripsQueryVariables = {
  filter?: ModelTripsFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListTripsQuery = {
  listTrips?:  {
    __typename: "ModelTripsConnection",
    items:  Array< {
      __typename: "Trips",
      id: string,
      trip_status?: TripStatus | null,
      pickup_lat?: string | null,
      pickup_lng?: string | null,
      dropoff_lat?: string | null,
      dropoff_lng?: string | null,
      pickup_location_description?: string | null,
      dropoff_location_description?: string | null,
      trip_cost?: number | null,
      TotalQtyItems?: number | null,
      receipt?: string | null,
      duration?: number | null,
      distance?: number | null,
      store_name?: string | null,
      store_image?: string | null,
      user_name?: string | null,
      user_photo?: string | null,
      mobile_number?: string | null,
      driver_name?: string | null,
      driver_photo?: string | null,
      driver_rating?: number | null,
      driver_mobile_number?: string | null,
      driver_car_model?: string | null,
      driver_car_make?: string | null,
      driver_car_color?: string | null,
      driver_car_plate_number?: string | null,
      sub_total?: number | null,
      redeem_points?: number | null,
      service_fee?: number | null,
      delivery_fee?: number | null,
      delivery_discount?: number | null,
      sum_total?: number | null,
      Driver?:  {
        __typename: "Driver",
        id: string,
        name: string,
        email: string,
        image?: string | null,
        phone_number?: string | null,
        rating?: number | null,
        inviteCode?: string | null,
        lng?: string | null,
        lat?: string | null,
        verificationStatus?: DriverRegistrationStatus | null,
        onlineStatus?: boolean | null,
        fcmToken?: string | null,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null,
      userID?: string | null,
      storeID: string,
      TripCarts?:  {
        __typename: "ModelTripCartConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Ratings?:  {
        __typename: "ModelRatingConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
      tripsDriverId?: string | null,
    } | null >,
    nextToken?: string | null,
    startedAt?: number | null,
  } | null,
};
// ...
export type GetDriverQueryVariables = {
  id: string,
};

export type GetDriverQuery = {
  getDriver?:  {
    __typename: "Driver",
    id: string,
    name: string,
    email: string,
    image?: string | null,
    phone_number?: string | null,
    rating?: number | null,
    inviteCode?: string | null,
    lng?: string | null,
    lat?: string | null,
    verificationStatus?: DriverRegistrationStatus | null,
    onlineStatus?: boolean | null,
    Cars?:  {
      __typename: "ModelCarConnection",
      items:  Array< {
        __typename: "Car",
        id: string,
        model: string,
        make: string,
        year: number,
        plate_number: string,
        color: string,
        driverID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    Ratings?:  {
      __typename: "ModelRatingConnection",
      items:  Array< {
        __typename: "Rating",
        id: string,
        rating?: number | null,
        comment?: string | null,
        driver_image?: string | null,
        driver_name?: string | null,
        customer_name?: string | null,
        customer_image?: string | null,
        tripsID: string,
        driverID: string,
        userID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    DriverWallets?:  {
      __typename: "ModelDriverWalletConnection",
      items:  Array< {
        __typename: "DriverWallet",
        id: string,
        amount: number,
        driverID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    LicenseDocuments?:  {
      __typename: "ModelLicenseDocumentConnection",
      items:  Array< {
        __typename: "LicenseDocument",
        id: string,
        image?: string | null,
        name?: string | null,
        expiry_date?: string | null,
        driverID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    RegistrationDocuments?:  {
      __typename: "ModelRegistrationDocumentConnection",
      items:  Array< {
        __typename: "RegistrationDocument",
        id: string,
        image?: string | null,
        name?: string | null,
        expiry_date?: string | null,
        driverID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    InsuranceDocuments?:  {
      __typename: "ModelInsuranceDocumentConnection",
      items:  Array< {
        __typename: "InsuranceDocument",
        id: string,
        image?: string | null,
        name?: string | null,
        expiry_date?: string | null,
        driverID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    fcmToken?: string | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};